Added Basic Leveling
Basic leveling has been added and test some. This implementation features:
Player XP Gain and Losses based on NPC Type, on Hurt and Kill Events were added. Some types are still missing from the config, but are weighted based on the type name; for instance 'mutantbear'.
Levels are stored, but still recalculated to check for level change. An event fires when the level changes.
Reporting now reports XP changes and can be controlled in the config.